requirements.txt
langdetect==1.0.9
pyyaml==6.0
pytest==7.3.1
requests==2.31.0
requests-mock==1.10.0
tenacity==8.2.2
